Crate bp_bridge_hub_cumulus

Primitives of all Cumulus-based bridge hubs.

Structs

AccountInfoStorageMapKeyProvider

Provides a storage key for account data.

BlockExecutionWeight

Importing a block with 0 Extrinsics.

BlockLength

Size limit of the Cumulus-based bridge hub blocks.

BlockWeights

Weight limit of the Cumulus-based bridge hub blocks.

BlockWeightsForAsyncBacking

Weight limit of the Cumulus-based bridge hub blocks when async backing is enabled.

ExtrinsicBaseWeight

Executing a NO-OP System::remarks Extrinsic.

Perbill

A fixed point representation of a number in the range [0, 1].

Constants

AVERAGE_BLOCK_INTERVAL

Average block interval in Cumulus-based parachains.

AVERAGE_ON_INITIALIZE_RATIO

All cumulus bridge hubs assume that about 5 percent of the block weight is consumed by on_initialize handlers. This is used to limit the maximal weight of a single extrinsic.

EXTRA_STORAGE_PROOF_SIZE

Number of extra bytes (excluding size of storage value itself) of storage proof, built at Polkadot-like chain. This mostly depends on number of entries in the storage trie. Some reserve is reserved to account future chain growth.

MAX_BRIDGE_HUB_HEADER_SIZE

Maximal bridge hub header size.

MAX_UNCONFIRMED_MESSAGES_IN_CONFIRMATION_TX

Maximal number of unconfirmed messages at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

MAX_UNREWARDED_RELAYERS_IN_CONFIRMATION_TX

Maximal number of unrewarded relayer entries at inbound lane for Cumulus-based parachains. Note: this value is security-relevant, decreasing it should not be done without careful analysis (like the one above).

NORMAL_DISPATCH_RATIO

All cumulus bridge hubs allow normal extrinsics to fill block up to 75 percent.

TX_EXTRA_BYTES

Maximal number of bytes, included in the signed Polkadot-like transaction apart from the encoded call itself.

Type Aliases

AccountId

Id of account on Polkadot-like chains.

AccountPublic

Public key of account on Polkadot-like chains.

AccountSigner

Public key of the chain account that may be used to verify signatures.

Address

The address format for describing accounts.

Balance

The balance of an account on Polkadot-like chain.

BlockNumber

Block number type used in Polkadot-like chains.

Hash

Hash type used in Polkadot-like chains.

Hasher

The type of object that can produce hashes on Polkadot-like chains.

Hashing

Hashing type.

Header

The header type used by Polkadot-like chains.

Nonce

Nonce of a transaction on the Polkadot-like chains.

Signature

Signature type used by Polkadot-like chains.

SignedBlock

Polkadot-like block signed with a Justification.

TransactionExtension

Signed extension that is used by all bridge hubs.

UncheckedExtrinsic

Unchecked Extrinsic type.